Move test_utils to d2
